move things into resource subdirectory of mapper
